pyproject.toml
[tool.poetry]
name = "doctr-api"
version = "0.1.2"
description = ""
authors = ["Kristian Mischke <kristianmischke@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
numpy = "^1.23.3"
python-doctr = {extras = ["torch"], version = "^0.6.0"}
weasyprint = "^56.1"
requests = "^2.28.1"
Pillow = "^9.2.0"
symspellpy = "^6.7.6"
pandas = "^1.5.0"
thefuzz = "^0.19.0"

torch = [
    {url = "https://download.pytorch.org/whl/cu117/torch-1.13.0%2Bcu117-cp310-cp310-linux_x86_64.whl", markers = "sys_platform == 'linux'"},
    {url = "https://download.pytorch.org/whl/cu117/torch-1.13.0%2Bcu117-cp310-cp310-win_amd64.whl", markers = "sys_platform == 'win32'"}
]
torchvision = [
    {url = "https://download.pytorch.org/whl/cu117/torchvision-0.14.0%2Bcu117-cp310-cp310-linux_x86_64.whl", markers = "sys_platform == 'linux'"},
    {url = "https://download.pytorch.org/whl/cu117/torchvision-0.14.0%2Bcu117-cp310-cp310-win_amd64.whl", markers = "sys_platform == 'win32'"}
]

[tool.poetry.dev-dependencies]
scikit-build = "^0.15.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.2.2"]
build-backend = "poetry.core.masonry.api"
